// Ecommerce/client/src/shared/models/basket.ts
export interface IBasketItem {
    quantity: number;
    imageFile: string;
    price: number;
    productId: string;
    productName: string;
}

export interface IBasket {
    userName: string;
    items: IBasketItem[];
    totalPrice: number;
}

export class Basket implements IBasket {
    userName: string = localStorage.getItem("basket_userName") || 'ahmed';
    totalPrice: number = 0;
    items: IBasketItem[] = [];
}

export interface IBasketTotal {
    total: number;
}